Rename files for new pcb designs using the RP2040
The files have been renamed so the sort better alphabetically and
they specify which build environment they are targeted for.

The scsi_accel files now specifically call out whether
they are for target mode or host mode (initiator mode).

The build environment is suffixed to their respective files.
Currently the only targets are RP2040 and BS2.
