+++ lib/ZuluSCSI_platform_RP2040/sdio.cpp
@@ -29,7 +29,7 @@
 // https://www.sdcard.org/downloads/pls/
 // "SDIO Physical Layer Simplified Specification Version 8.00"
 
-#include "rp2040_sdio.h"
+#include "sdio.h"
 #include <hardware/pio.h>
 #include <hardware/dma.h>
 #include <hardware/gpio.h>
@@ -37,9 +37,9 @@
 #include <ZuluSCSI_log.h>
 
 #ifdef ZULUSCSI_BS2
-#include "rp2040_sdio_BS2.pio.h"
+#include "sdio_BS2.pio.h"
 #else
-#include "rp2040_sdio.pio.h"
+#include "sdio_RP2040.pio.h"
 #endif
 
 #define SDIO_PIO pio1
